fix(protocol): pack enum fields at their declared wire width
The pack/unpack codegen hardcoded uint8 for every enum field, ignoring the
field's size_bytes. Upstream defines ButtonGesture and ButtonTargetType as
size_bytes: 2 (the only 2-byte enums), so ButtonAction packed to 18 bytes
instead of 20 and Button.State's payload to 731 instead of 811 -- a real
LIFX LAN client would misparse Button by 80 bytes. Round-trips passed only
because pack and unpack were symmetrically wrong.

- generator: derive enum wire width from the field's size_bytes
  (1->uint8, 2->uint16, 4->uint32) via new enum_wire_type helper
- generator: apply_field_overrides now fails loud if an override key is
  absent upstream, instead of silently shipping the generic layout
- button_handlers: normalise each padded button to exactly 5 actions so
  Button.State always round-trips through pack()/unpack()
- regenerate protocol_types.py: ButtonAction gesture/target_type now uint16

Wire sizes now correct: ButtonAction=20, Button=101, Button.State payload=811.

+++ packages/lifx-emulator-core/src/lifx_emulator/protocol/protocol_types.py
@@ -216,9 +216,9 @@ class ButtonAction:
         result = b""
 
         # gesture: ButtonGesture (enum)
-        result += serializer.pack_value(int(self.gesture), "uint8")
+        result += serializer.pack_value(int(self.gesture), "uint16")
         # target_type: ButtonTargetType (enum)
-        result += serializer.pack_value(int(self.target_type), "uint8")
+        result += serializer.pack_value(int(self.target_type), "uint16")
         # target: ButtonTarget
         result += self.target.pack()
 
@@ -232,12 +232,12 @@ class ButtonAction:
         current_offset = offset
         # gesture: ButtonGesture (enum)
         gesture_raw, current_offset = serializer.unpack_value(
-            data, "uint8", current_offset
+            data, "uint16", current_offset
         )
         gesture = ButtonGesture(gesture_raw)
         # target_type: ButtonTargetType (enum)
         target_type_raw, current_offset = serializer.unpack_value(
-            data, "uint8", current_offset
+            data, "uint16", current_offset
         )
         target_type = ButtonTargetType(target_type_raw)
         # target: ButtonTarget
